Remove class Node and insert in class HandleLinkedHashMap
