+++ src/classes/Epic.java
@@ -4,11 +4,10 @@ import classes.enums.Class;
 
 import java.util.HashMap;
 
-public class Epic extends Task {
-
+public class Epic extends Task{
     private HashMap<Integer, Subtask> epicSubtasks;
 
-     public Epic(String title, String description) {
+     public Epic(String title, String description){
          super(title, description);
          epicSubtasks =  new HashMap<>();
      }
@@ -21,13 +20,12 @@ public class Epic extends Task {
      public HashMap<Integer, Subtask> getSubMap(){
          return epicSubtasks;
      }
-
      public Subtask getSubTask(int id){
          return epicSubtasks.get(id);
      }
 
     @Override
-    public String toString() {
+    public String toString(){
         return "Epic: " + this.getTitle() + " [id:" + this.getId() + ", status:" + this.getStatus() + ", number of subs:" + getSubMap().size() + "]";
     }
 }
