--- src/classes/Node.java
@@ -1,29 +1,0 @@
-package classes;
-
-import java.util.Objects;
-
-public class Node {
-
-    public Task data;
-    public Node next;
-    public Node prev;
-
-    public Node(Node prev, Task data, Node next) {
-        this.data = data;
-        this.next = next;
-        this.prev = prev;
-    }
-
-    @Override
-    public boolean equals(Object o) {
-        if(this == o) return true;
-        if(o == null || this.getClass() != o.getClass()) return false;
-        Node node = (Node) o;
-        return data.equals(node.data);
-    }
-
-    @Override
-    public int hashCode() {
-        return 31 * Objects.hashCode(data);
-    }
-}
