+++ src/classes/Subtask.java
@@ -5,7 +5,6 @@ import classes.enums.Class;
 public class Subtask extends Task{
     private int motherId;
     public Subtask(String title, String description) {
-
         super(title, description);
     }
 
@@ -23,7 +22,7 @@ public class Subtask extends Task{
     }
 
     @Override
-    public String toString() {
+    public String toString(){
         return "Subtask: " + this.getTitle() + " [id: " + this.getId() + ", status: " + this.getStatus() + "mother epic id:" + motherId + "]";
     }
