+++ src/controllers/InMemoryHistoryManager.java
@@ -5,8 +5,6 @@ import controllers.interfaces.HistoryManager;
 
 import java.util.*;
 
-import classes.Node;
-
 public class InMemoryHistoryManager implements HistoryManager {
     private HandleLinkedHashMap handleMap;
 
@@ -32,6 +30,31 @@ public class InMemoryHistoryManager implements HistoryManager {
 }
 
 class HandleLinkedHashMap {
+    public class Node {
+
+        public Task data;
+        public Node next;
+        public Node prev;
+
+        public Node(Node prev, Task data, Node next) {
+            this.data = data;
+            this.next = next;
+            this.prev = prev;
+        }
+
+        @Override
+        public boolean equals(Object o) {
+            if(this == o) return true;
+            if(o == null || this.getClass() != o.getClass()) return false;
+            Node node = (Node) o;
+            return data.equals(node.data);
+        }
+
+        @Override
+        public int hashCode() {
+            return 31 * Objects.hashCode(data);
+        }
+    }
 
     Map<Integer, Node> handleLinkedMap;
     private Node head;
